ipa: rpi: pisp: Enable decompand algorithm based on bit depth
This patch adds support for a "bitdepth" parameter in the tuning JSON for the decompand block.
When specified (e.g. "bitdepth": 12), the decompand algorithm will only be enabled if the camera's current bit depth matches the configured value.
For all other bit depths, the algorithm is disabled.

If the "bitdepth" parameter is not present in the JSON, which means the decompand algorithm is always enabled.

Signed-off-by: Sena Asotani <aso.fam429@gmail.com>